generative_ai/observability/openllmetry_integration/requirements.txt
# Minimal packages to use OpenLLMetry w/GCP Cloud O11y
opentelemetry-exporter-gcp-trace==1.7.0
traceloop-sdk==0.33.12

# Package used to provide "_logging_exporter" to additionally
# integrate with Cloud Logging in addition to Cloud Trace.
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2

# Package used to integrate with Cloud Monitoring.
opentelemetry-exporter-gcp-monitoring==1.7.a0

# -- DEMO SPECIFIC PACKAGES --
# The following packages are necessary for the
# particular demo involving langchain, but may not
# necessarily be required for use with other framworks.

# For langchain v0.3
# From https://python.langchain.com/v0.2/docs/versions/overview/
# With updates based on https://python.langchain.com/docs/versions/v0_3/
langchain = ">=0.3,<0.4"
langchain-core = ">=0.3,<0.4"
langchain-text-splitters = ">=0.3,<0.4"
pydantic = ">=2,<3"
SQLAlchemy = ">=1.4,<3"
requests = "^2"
PyYAML = ">=5.3"
numpy = "^1"
aiohttp = "^3.8.3"
tenacity = "^8.1.0"
jsonpatch = "^1.33"

# For using LangChain with VertexAI as the backend.
langchain-google-vertexai = ">=2,<3"
